requirements.txt
celery
flask
flask_sqlalchemy
psycopg2
redis
requests
tornado==4.5.3
pynsq
